chore(wizard-refactor): extract resolvers/course-by-name
ResolvedPlaybook + CourseResolution + resolveCourseByName.
Dispatcher 11/11 + lint 17/17 green.
